Fix a "None" in normal cringo title
The cringo start embed looks at this file for the name to call the kind of cringo that just started. For 4-cringo, which is normal cringo, this was returning None and making the embed say "NoneCRINGO". This can be fixed by adding a cheeky zero width space in place of an empty value for 4-cringo's name. Just hope all the clients render it correctly 🙂